+++ NEWS.md
@@ -5,9 +5,12 @@ main
 ------
 <!--- # Add changes since the most recent release here --->
 
+- Bugfix; fixed the evaporation scheme of SB2006 to prevent returning NaN values when limiters are not applied. ([#420](https://github.com/CliMA/CloudMicrophysics.jl/pull/415))
 
 v0.20.0
 ------
+- Added modified terminal velocity without limiters for SB2006 ([#415](https://github.com/CliMA/CloudMicrophysics.jl/pull/415))
+
 - Add lognormal droplet distribution in parcel model ([#414](https://github.com/CliMA/CloudMicrophysics.jl/pull/414))
 
 - Added calling function to grab artifacts ([#413](https://github.com/CliMA/CloudMicrophysics.jl/pull/413))